+++ src/firecracker/src/main.rs
@@ -3,7 +3,7 @@
 mod api_server_adapter;
 mod metrics;
 
-use std::fs;
+use std::fs::{self, File};
 use std::io;
 use std::panic;
 use std::path::PathBuf;
@@ -12,13 +12,12 @@ use std::sync::{Arc, Mutex};
 
 use logger::{error, info, IncMetric, ProcessTimeReporter, LOGGER, METRICS};
 use polly::event_manager::EventManager;
-use seccomp::{BpfProgram, SeccompLevel};
+use seccomp::BpfThreadMap;
 use snapshot::Snapshot;
-use std::fs::File;
 use utils::arg_parser::{ArgParser, Argument};
 use utils::terminal::Terminal;
 use utils::validators::validate_instance_id;
-use vmm::default_syscalls::get_seccomp_filter;
+use vmm::default_syscalls::{get_custom_filters, get_default_filters, get_empty_filters};
 use vmm::resources::VmResources;
 use vmm::signal_handler::{mask_handled_signals, SignalManager};
 use vmm::version_map::{FC_VERSION_TO_SNAP_VERSION, VERSION_MAP};
@@ -77,12 +76,10 @@ fn main() {
                 .help("MicroVM unique identifier."),
         )
         .arg(
-            Argument::new("seccomp-level")
+            Argument::new("seccomp-filter")
                 .takes_value(true)
-                .default_value("2")
                 .help(
-                    "Level of seccomp filtering (0: no filter | 1: filter by syscall number | 2: filter by syscall \
-                     number and argument values) that will be passed to executed path as argument."
+                    "Optional parameter which allows specifying the path to a custom seccomp filter. For advanced users."
                 ),
         )
         .arg(
@@ -110,6 +107,12 @@ fn main() {
                 .takes_value(false)
                 .requires("config-file")
                 .help("Optional parameter which allows starting and using a microVM without an active API socket.")
+        )
+        .arg(
+            Argument::new("no-seccomp")
+                .takes_value(false)
+                .help("Optional parameter which allows starting and using a microVM without seccomp filtering. \
+                    Not recommended.")
         )
         .arg(
             Argument::new("log-path")
@@ -221,16 +224,14 @@ fn main() {
         });
     }
 
-    // It's safe to unwrap here because the field's been provided with a default value.
-    let seccomp_level = arguments.single_value("seccomp-level").unwrap();
-    let seccomp_filter = get_seccomp_filter(
-        SeccompLevel::from_string(&seccomp_level).unwrap_or_else(|err| {
-            panic!("Invalid value for seccomp-level: {}", err);
-        }),
-    )
-    .unwrap_or_else(|err| {
-        panic!("Could not create seccomp filter: {}", err);
-    });
+    let mut seccomp_filters = match arguments.flag_present("no-seccomp") {
+        true => get_empty_filters(),
+        false => match arguments.single_value("seccomp-filter") {
+            Some(path) => get_custom_filters(File::open(&path).expect("Seccomp file error: {}"))
+                .expect("Could not create seccomp filter"),
+            None => get_default_filters().expect("Could not create seccomp filter"),
+        },
+    };
 
     let vmm_config_json = arguments
         .single_value("config-file")
@@ -264,7 +265,7 @@ fn main() {
         let process_time_reporter =
             ProcessTimeReporter::new(start_time_us, start_time_cpu_us, parent_cpu_time_us);
         api_server_adapter::run_with_api(
-            seccomp_filter,
+            &mut seccomp_filters,
             vmm_config_json,
             bind_path,
             instance_info,
@@ -272,8 +273,12 @@ fn main() {
             boot_timer_enabled,
         );
     } else {
+        let mut seccomp_filters: BpfThreadMap = seccomp_filters
+            .into_iter()
+            .filter(|(k, _)| k != "api")
+            .collect();
         run_without_api(
-            seccomp_filter,
+            &mut seccomp_filters,
             vmm_config_json,
             &instance_info,
             boot_timer_enabled,
@@ -333,7 +338,7 @@ fn print_snapshot_data_format(snapshot_path: &str) {
 
 // Configure and start a microVM as described by the command-line JSON.
 fn build_microvm_from_json(
-    seccomp_filter: BpfProgram,
+    seccomp_filters: &mut BpfThreadMap,
     event_manager: &mut EventManager,
     config_json: String,
     instance_info: &InstanceInfo,
@@ -348,7 +353,7 @@ fn build_microvm_from_json(
             process::exit(i32::from(vmm::FC_EXIT_CODE_BAD_CONFIGURATION));
         });
     vm_resources.boot_timer = boot_timer_enabled;
-    let vmm = vmm::builder::build_microvm_for_boot(&vm_resources, event_manager, &seccomp_filter)
+    let vmm = vmm::builder::build_microvm_for_boot(&vm_resources, event_manager, seccomp_filters)
         .unwrap_or_else(|err| {
             error!(
                 "Building VMM configured from cmdline json failed: {:?}",
@@ -362,7 +367,7 @@ fn build_microvm_from_json(
 }
 
 fn run_without_api(
-    seccomp_filter: BpfProgram,
+    seccomp_filters: &mut BpfThreadMap,
     config_json: Option<String>,
     instance_info: &InstanceInfo,
     bool_timer_enabled: bool,
@@ -391,7 +396,7 @@ fn run_without_api(
     // - VmResources is not used without api,
     // - An `Arc` reference of the built `Vmm` is plugged in the `EventManager` by the builder.
     build_microvm_from_json(
-        seccomp_filter,
+        seccomp_filters,
         &mut event_manager,
         // Safe to unwrap since '--no-api' requires this to be set.
         config_json.unwrap(),