+++ src/vmm/src/persist.rs
@@ -28,7 +28,7 @@ use cpuid::common::{get_vendor_id_from_cpuid, get_vendor_id_from_host};
 use arch::regs::{get_manufacturer_id_from_host, get_manufacturer_id_from_state};
 use logger::{error, info};
 use polly::event_manager::EventManager;
-use seccomp::BpfProgramRef;
+use seccomp::BpfThreadMap;
 use snapshot::Snapshot;
 use versionize::{VersionMap, Versionize, VersionizeResult};
 use versionize_derive::Versionize;
@@ -389,7 +389,7 @@ pub fn snapshot_state_sanity_check(
 /// Loads a Microvm snapshot producing a 'paused' Microvm.
 pub fn restore_from_snapshot(
     event_manager: &mut EventManager,
-    seccomp_filter: BpfProgramRef,
+    seccomp_filters: &mut BpfThreadMap,
     params: &LoadSnapshotParams,
     version_map: VersionMap,
 ) -> std::result::Result<Arc<Mutex<Vmm>>, LoadSnapshotError> {
@@ -410,7 +410,7 @@ pub fn restore_from_snapshot(
         microvm_state,
         guest_memory,
         track_dirty_pages,
-        seccomp_filter,
+        seccomp_filters,
     )
     .map_err(BuildMicroVm)
 }