+++ src/vmm/tests/integration_tests.rs
@@ -6,12 +6,12 @@ use std::thread;
 use std::time::Duration;
 
 use polly::event_manager::EventManager;
-use seccomp::{BpfProgram, SeccompLevel};
+use seccomp::{BpfProgram, BpfThreadMap};
 use snapshot::Snapshot;
 use utils::tempfile::TempFile;
 use vmm::builder::build_microvm_from_snapshot;
 use vmm::builder::{build_microvm_for_boot, setup_serial_device};
-use vmm::default_syscalls::get_seccomp_filter;
+use vmm::default_syscalls::get_empty_filters;
 use vmm::persist;
 use vmm::persist::{snapshot_state_sanity_check, LoadSnapshotError, MicrovmState};
 use vmm::resources::VmResources;
@@ -49,9 +49,10 @@ fn test_build_microvm() {
     {
         let resources: VmResources = MockVmResources::new().into();
         let mut event_manager = EventManager::new().unwrap();
-        let empty_seccomp_filter = get_seccomp_filter(SeccompLevel::None).unwrap();
+        let mut empty_seccomp_filters = get_empty_filters();
 
-        let vmm_ret = build_microvm_for_boot(&resources, &mut event_manager, &empty_seccomp_filter);
+        let vmm_ret =
+            build_microvm_for_boot(&resources, &mut event_manager, &mut empty_seccomp_filters);
         assert_eq!(format!("{:?}", vmm_ret.err()), "Some(MissingKernelConfig)");
     }
 
@@ -100,7 +101,10 @@ fn test_vmm_seccomp() {
 
             // The customer "forgot" to allow the KVM_RUN ioctl.
             let filter: BpfProgram = MockSeccomp::new().without_kvm_run().into();
-            let vmm = build_microvm_for_boot(&resources, &mut event_manager, &filter).unwrap();
+            let mut filters = BpfThreadMap::new();
+            filters.insert("vmm".to_string(), filter.clone());
+            filters.insert("vcpu".to_string(), filter);
+            let vmm = build_microvm_for_boot(&resources, &mut event_manager, &mut filters).unwrap();
             // Give the vCPUs a chance to attempt KVM_RUN.
             thread::sleep(Duration::from_millis(200));
             // Should never get here.
@@ -347,7 +351,7 @@ fn verify_load_snapshot(snapshot_file: TempFile, memory_file: TempFile) {
         0 => {
             set_panic_hook();
             let mut event_manager = EventManager::new().unwrap();
-            let empty_seccomp_filter = get_seccomp_filter(SeccompLevel::None).unwrap();
+            let mut empty_seccomp_filters = get_empty_filters();
 
             // Deserialize microVM state.
             let snapshot_file_metadata = snapshot_file.as_file().metadata().unwrap();
@@ -369,7 +373,7 @@ fn verify_load_snapshot(snapshot_file: TempFile, memory_file: TempFile) {
                 microvm_state,
                 mem,
                 false,
-                &empty_seccomp_filter,
+                &mut empty_seccomp_filters,
             )
             .unwrap();
             // For now we're happy we got this far, we don't test what the guest is actually doing.